myproject/requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.1
djangorestframework==3.15.2
django-filter==2.0.0
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1